Struct cfrp::primitives::Builder

pub struct Builder {
    // some fields omitted
}

Builder is used to construct topologies.

Basic builder pattern - Topology::build accepts a function which takes a state type T and a mutable builder. The builder can be used to create Channels and to add nodes to the topology

Methods

impl Builder

fn new() -> Builder

Create a new Builder

fn add<A, SA>(&self, root: SA) -> Branch<A> where SA: 'static + Signal<A>, A: 'static + Clone + Send

Add a signal to the topology

Returns a Branch<A>, allowing root to be used as input more than once

Example

use cfrp::*;
use cfrp::primitives::*;

let b = Builder::new();
 
// Topologies only execute transformations which have been added to a builder.
let fork = b.add(b.value(1).lift(|i| { i + 1} ));

// `add` returns a signal that can be used more than once
b.add(fork.clone().lift(|i| { i - 1 } ));
b.add(fork.lift(|i| { -i }));

fn listen<A, T>(&self, initial: A, input: T) -> Channel<A> where T: 'static + Input<A> + Send, A: 'static + Clone + Send

Listen to input and push received data into the topology

All data must enter the topology via a call to listen; this function ensures data syncronization across the topology. Each listener runs in its own thread

Example

use std::sync::mpsc::*;
use cfrp::*;
use cfrp::primitives::*;

let b = Builder::new();
 
let (tx, rx): (Sender<usize>, Receiver<usize>) = channel();

// Receive data on `rx` and expose it as a signal with initial value 
//`initial`.  This is necessary because the topology must maintain 
// consistency between threads, so any message sent to any input is 
// propagated to all other inputs as "no-change" messages.
let signal = b.listen(0, rx);

fn async<A, SA>(&self, root: SA) -> Channel<A> where SA: 'static + Signal<A>, A: 'static + Clone + Send

Combination of adding a signal and a channel

Async allows signals to be processed downstream out of order. Internally, the output of root is sent to new input channel. The result is that long-running processes can be handled outside of the synchronized topology process, and the result can be handled when it's available.

use std::sync::mpsc::*;
use cfrp::*;
use cfrp::primitives::*;

let b = Builder::new();
 
// This will now happen without blocking the rest of the topology
let result = b.async(
    b.value(0).fold(0, |i, j| {
        // Some very expensive code in here...
    })
);

// ...and `result` will receive the output value when it's done
b.add(
    b.value(0).lift2(result, |i, j| { (i, j) })
);

fn value<T>(&self, v: T) -> Channel<T> where T: 'static + Clone + Send

Creats a channel with constant value v