Trait cfrp::Lift

pub trait Lift<A>: Signal<A> + Sized {
    fn lift<F, B>(self, f: F) -> LiftSignal<F, A, B> where Self: 'static, F: 'static + Send + Fn(A) -> B, A: 'static + Send, B: 'static + Send { ... }
}

Apply a pure function F to a data source Signal<A>, generating a transformed output data source Signal<B>.

Other names for this operation include "map" or "collect". f will be run in self's thread

Because F is assumed to be pure, it will only be evaluated for new data that has changed since the last observation. If side-effects are desired, use fold instead.

Example

use cfrp::*;
use cfrp::primitives::*;

// Create a constant-valued channel with value 1 and return a channel with
// value 2.  The value will only be calculated with `i` changes (which in 
// this case is never)
Builder::new().value(1)
    .lift(|i| { i + 1 });

Provided Methods

fn lift<F, B>(self, f: F) -> LiftSignal<F, A, B> where Self: 'static, F: 'static + Send + Fn(A) -> B, A: 'static + Send, B: 'static + Send